requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.0
Django==4.0.3
django-livereload-server==0.4
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
tornado==6.1
whitenoise==6.0.0
